Add arguments vector to execute function (#5)
* add monetizer

* new monetizer implementation

* new apply_parameter setup with impls

* add apply_parameters to execute

* implement support for i32

* update README

* Convert integer types to SQLparameters

* Rename SQLParameters to SQLparameter (without s)

* Use vectors instead of slices for now

Vectors are dynamic, i.e. can be built at runtime. Slices must be
compiled or converted from vectors.

Also add some glue code (monetizer::to_sqlparameter) to make parameter
creation a bit easier.

* apply_parameters does not need to return Option

It always returns a string.

* Fix a small bug in apply_parameters

We need to push the substring to the result even if it does not contain '{}'.

* Fix the tests

Co-authored-by: Panagiotis Koutsourakis <kutsurak@monetdbsolutions.com>